+++ src/pages/jj/Main/Main.scss
@@ -83,8 +83,18 @@
             }
 
             .mainImg {
+              position: relative;
               width: 100%;
               height: 60%;
+
+              .bigHeart {
+                position: absolute;
+                top: 50%;
+                left: 50%;
+                transform: translate(-50%, -50%);
+                color: rgba(255, 6, 6, 0.603);
+                font-size: 8rem;
+              }
 
               .mainFeedImg {
                 width: 100%;
