+++ src/pages/jj/config/api.js
@@ -1,0 +1,3 @@
+const APIKEY = 'http://localhost:3000/data/jj/Data.json';
+
+export default APIKEY;
